+++ CK.MQTT.Common/OutgoingPackets/FromMemoryOutgoingPacket.cs
@@ -9,9 +9,9 @@ namespace CK.MQTT.Common.OutgoingPackets
     public class FromMemoryOutgoingPacket : IOutgoingPacket
     {
         readonly ReadOnlyMemory<byte> _readOnlyMemory;
-        readonly int _packetId;
+        readonly uint _packetId;
 
-        public FromMemoryOutgoingPacket( ReadOnlyMemory<byte> readOnlyMemory, QualityOfService qos, int packetId )
+        public FromMemoryOutgoingPacket( ReadOnlyMemory<byte> readOnlyMemory, QualityOfService qos, uint packetId )
         {
             Debug.Assert( readOnlyMemory.Length > 0 );
             _readOnlyMemory = readOnlyMemory;
@@ -20,9 +20,9 @@ namespace CK.MQTT.Common.OutgoingPackets
         }
 
         public QualityOfService Qos { get; }
-        public int PacketId { get => _packetId; set => throw new NotSupportedException(); }
+        public uint PacketId { get => _packetId; set => throw new NotSupportedException(); }
 
-        public int GetSize( ProtocolLevel protocolLevel ) => _readOnlyMemory.Length;
+        public uint GetSize( ProtocolLevel protocolLevel ) => (uint)_readOnlyMemory.Length;
 
         public async ValueTask<IOutgoingPacket.WriteResult> WriteAsync( ProtocolLevel protocolLevel, PipeWriter writer, CancellationToken cancellationToken )
         {
